moved HalCrawler and MockMvcHalResourceLoaderConfiguration into testing module
